+++ src/views/explainPlan/ExplainPlanPanel.ts
@@ -179,6 +179,7 @@ export class ExplainPlanPanel {
 
             html = html.replace('{{CSS_URI}}', cssUri.toString());
             html = html.replace('{{JS_URI}}', jsUri.toString());
+            html = html.replace(/\{\{CSP_SOURCE\}\}/g, this._panel.webview.cspSource);
 
             return html;
         } catch (error) {