+++ src/views/tableDesigner/TableDesignerPanel.ts
@@ -323,6 +323,7 @@ export class TableDesignerPanel {
 
             html = html.replace('{{CSS_URI}}', cssUri.toString());
             html = html.replace('{{JS_URI}}', jsUri.toString());
+            html = html.replace(/\{\{CSP_SOURCE\}\}/g, this._panel.webview.cspSource);
 
             const configData = {
                 mode: this._mode,